login add

+++ app/Http/Controllers/Event/DeleteController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Event;
 use App\Models\Event_DateTime;
+use App\Services\EventService;
+use Symfony\Component\HttpKernel\Exception\AccessdeniedHttpException;
 
 class DeleteController extends Controller
 {
@@ -15,9 +17,13 @@ class DeleteController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function __invoke(Request $request)
-    {
+    public function __invoke(Request $request, EventService $eventService)
+    {   
+        
         $eventId = (int) $request->route('eventId');
+        if (!$eventService->checkOwnEvent($request->user()->id,$eventId)) {
+            throw new AccessDeniedHttpException();
+        }
         $event = Event::where('id',$eventId)->firstOrFail();
         $event->event_datetimes()->delete();
         $event->delete();
